refactor: rework layout for recvpacket rechecktx
